+++ packages/components/src/ComboBox/ComboBox.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 
-import { useButton } from '@react-aria/button';
 import { useComboBox } from '@react-aria/combobox';
 import { useFilter } from '@react-aria/i18n';
 
@@ -82,7 +81,6 @@ export const ComboBox = ({
 
   // TODO: until `react-aria` gives us error and description props.
   const errorMessageProps = { 'aria-invalid': error };
-  const { buttonProps } = useButton(triggerProps, buttonRef);
   const { label, description, errorMessage } = props;
 
   return (
@@ -103,7 +101,7 @@ export const ComboBox = ({
             <Button
               className="absolute right-2 h-4 w-4 border-none bg-transparent p-0"
               ref={buttonRef}
-              {...buttonProps}
+              {...triggerProps}
             >
               <ChevronDown className="h-4 w-4" />
             </Button>